requirements.txt
pybit==5.11.0
requests==2.32.5
python-dotenv==1.1.1
